run: move monitor logic out of run
